Move context hooks into /src/hooks/contexts, create new useDashboardContext hook, and update components using the ColorContext to point to the right file in their imports
