+++ src/contexts/dashboardContext.js
@@ -1,0 +1,7 @@
+import { createContext, useState } from 'react'
+import { useCookies } from 'react-cookie'
+import PropTypes from 'prop-types'
+
+const DashboardContext = createContext()
+
+export { DashboardContext }
